Fix to some of the rank icon rendering issues.
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@3767 3db68df2-c116-0410-a063-a993310a9797

+++ source/simulation/BaseEntity.h
@@ -84,7 +84,7 @@ public:
 
 	//Rank properties
 	float m_rankWidth;
-	int m_rankHeight;
+	float m_rankHeight;
 	CStr m_rankName;
 
 	// Minimap properties
